Revert rev 468.  This caused a regression.
This change assumed that the makeopts files in /etc/or ~/ were full makeopts files.
For example, if you had a makeopts file in ~/ that only had one line to ensure
that you always built with DEBUG_THREADS enabled, it would no longer work, because
the menuselect sanity check would freak out.